bonuses. added Withdraw handler

+++ internal/bonuses/storage/managers/interface.go
@@ -7,9 +7,9 @@ import (
 )
 
 type BaseBonusesManager interface {
-	AddBonuses(ctx context.Context, userID int64, count int64) error
-	GetBonuses(ctx context.Context, userID int64) (int64, error)
-	WithdrawBonuses(ctx context.Context, userID int64, count int64) (bool, error)
-	GetWithdrawnBonuses(ctx context.Context, userID int64) (int64, error)
+	AddBonuses(ctx context.Context, userID int64, count float32) error
+	GetBonuses(ctx context.Context, userID int64) (float32, error)
+	WithdrawBonuses(ctx context.Context, userID int64, count float32) (bool, error)
+	GetWithdrawnBonuses(ctx context.Context, userID int64) (float32, error)
 	GetWithdrawals(ctx context.Context, userID int64) ([]data.Withdrawal, error)
 }
