+++ internal/bonuses/storage/managers/postgresql/manager.go
@@ -83,19 +83,19 @@ func (p *postgresDB) Close() error {
 	return p.database.Close()
 }
 
-func (p *postgresDB) AddBonuses(ctx context.Context, userID int64, count int64) error {
+func (p *postgresDB) AddBonuses(ctx context.Context, userID int64, count float32) error {
 	return nil
 }
 
-func (p *postgresDB) GetBonuses(ctx context.Context, userID int64) (int64, error) {
+func (p *postgresDB) GetBonuses(ctx context.Context, userID int64) (float32, error) {
 	return -1, nil
 }
 
-func (p *postgresDB) WithdrawBonuses(ctx context.Context, userID int64, count int64) (bool, error) {
+func (p *postgresDB) WithdrawBonuses(ctx context.Context, userID int64, count float32) (bool, error) {
 	return false, nil
 }
 
-func (p *postgresDB) GetWithdrawnBonuses(ctx context.Context, userID int64) (int64, error) {
+func (p *postgresDB) GetWithdrawnBonuses(ctx context.Context, userID int64) (float32, error) {
 	return -1, nil
 }
